backend/requirements.txt
# FastAPI y servidor
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Base de datos
sqlalchemy==2.0.23
pymysql==1.1.0
mysqlclient==2.2.0

# Autenticación y seguridad
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1

# Validación
pydantic[email]==2.5.0
email-validator==2.1.0

# Utilidades
python-dotenv==1.0.0
